feat: properly index and search datasets and jobs
