+++ api/src/main/java/marquez/api/v2beta/SearchResource.java
@@ -48,7 +48,7 @@ public class SearchResource {
   @Produces(APPLICATION_JSON)
   @Path("jobs")
   public Response searchJobs(@QueryParam("q") @NotBlank String query) throws IOException {
-    if (searchService.isEnabled()) {
+    if (!searchService.isEnabled()) {
       return Response.status(Response.Status.SERVICE_UNAVAILABLE).build();
     }
     return formatOpenSearchResponse(this.searchService.searchJobs(query));
@@ -61,7 +61,7 @@ public class SearchResource {
   @Produces(APPLICATION_JSON)
   @Path("datasets")
   public Response searchDatasets(@QueryParam("q") @NotBlank String query) throws IOException {
-    if (searchService.isEnabled()) {
+    if (!searchService.isEnabled()) {
       return Response.status(Response.Status.SERVICE_UNAVAILABLE).build();
     }
     return formatOpenSearchResponse(this.searchService.searchDatasets(query));
